requirements.txt
gradio[oauth]==4.44.1
chromadb==0.5.5
sentence-transformers==3.0.1
pypdf==4.3.1
python-pptx==1.0.2
beautifulsoup4==4.12.3
requests==2.32.3
gTTS==2.5.3
huggingface_hub>=0.31.0,<1.0

